net: shell: Add conn command
Move "conn" command to a separate file.
