+++ subsys/net/lib/shell/CMakeLists.txt
@@ -9,6 +9,7 @@ zephyr_library_link_libraries_ifdef(CONFIG_MBEDTLS mbedTLS)
 zephyr_library_sources(allocs.c)
 zephyr_library_sources(arp.c)
 zephyr_library_sources(capture.c)
+zephyr_library_sources(conn.c)
 
 zephyr_library_include_directories(
   ${ZEPHYR_BASE}/kernel/include
